+++ src/sapphire_config/metadata.py
@@ -1,3 +1,0 @@
-"""Package metadata"""
-
-__version__ = "1.0.4"